Todo listo y andando

+++ BackEndExchange/Controladores/CompraController.cs
@@ -61,12 +61,12 @@ namespace BackEndExchange.Controladores
         CriptomonedaModel cm = new CriptomonedaModel();
         cm.IdCriptomoneda = c.IdCriptomoneda;
         cm.Nombre = c.Nombre.Trim();
-        cm.PrecioCompra = (double)c.PrecioCompra;
-        cm.StockDisponible = (double)c.StockDisponible;
+        cm.PrecioCompra = (decimal)c.PrecioCompra;
+        cm.StockDisponible = (decimal)c.StockDisponible;
         cm.Simbolo = c.Simbolo.Trim();
-        cm.StockTotal = (double)c.StockTotal;
+        cm.StockTotal = (decimal)c.StockTotal;
         cm.ImagenUrl = c.ImagenUrl.Trim();
-        cm.PorcentajeGanancia = (double)c.PorcentajeGanancia;
+        cm.PorcentajeGanancia = (decimal)c.PorcentajeGanancia;
 
         TipoMovimientoModel tmm = new TipoMovimientoModel();
         tmm.IdTiposMovimiento = movimiento.IdTiposMovimiento;
@@ -78,7 +78,7 @@ namespace BackEndExchange.Controladores
         ConfiguracionCompraModel ccm = new ConfiguracionCompraModel();
         ccm.IdCriptomoneda = cm.IdCriptomoneda;
         ccm.NombreCriptomoneda = cm.Nombre;
-        ccm.PrecioVenta = getPrecioVentaPesos(cm.PrecioCompra, cm.PorcentajeGanancia);
+        ccm.PrecioVenta = getPrecioVentaPesos((double)cm.PrecioCompra, (double)cm.PorcentajeGanancia);
         ccm.Comision = tmm.Comision;
         ccm.IdTipoMovimiento = tmm.IdTiposMovimiento;
