+++ BackEndExchange/Model/Request/CriptomonedaModel.cs
@@ -9,11 +9,11 @@ namespace BackEndExchange.Model.Request
     {
     public int IdCriptomoneda { get; set; }
     public string Nombre { get; set; }
-    public double PrecioCompra { get; set; }
-    public double StockDisponible { get; set; }
+    public decimal PrecioCompra { get; set; }
+    public decimal StockDisponible { get; set; }
     public string Simbolo { get; set; }
-    public double StockTotal { get; set; }
-    public double PorcentajeGanancia { get; set; }
+    public decimal StockTotal { get; set; }
+    public decimal PorcentajeGanancia { get; set; }
     public string ImagenUrl { get; set; }
     public DateTime? FechaBaja { get; set; }
   }
